+++ CHANGES.md
@@ -1,5 +1,13 @@
 # Change Log
 
+### 1.124 - 2024-12-01
+
+#### @cesium/engine
+
+##### Additions :tada:
+
+- Added `Entity.trackingReferenceFrame` property to allow tracking entities in their own inertial reference frame. [#12194](https://github.com/CesiumGS/cesium/pull/12194)
+
 ### 1.123.1 - 2024-11-07
 
 #### @cesium/engine